Add a file transform to jest to handle assets (#392)
closes #387

Note that the issue mentions adding an asset import to a story in
`react-sdk`. Right now you also get type errors there on those. In the
designer app you don't get those because types for assets are declared
within the `@remix-run/dev` typedefs that you reference:

https://github.com/webstudio-is/webstudio-designer/blob/f9b625b3c31624f92d4d5a7d4235ba703ee28fcd/apps/designer/remix.env.d.ts#L1

This is a separate problem though and should be handled separately.