Search tree shift by selector (#1678)
Here refactored horizontal shift in tree to search available placement
by selector. This enables us to add restricting ancestors with
requiredAncestorss and invalidAncestors options in meta (in the next
PR).

The idea of this refactoring to do similar search for both left and
right directions based on item selectors which means from deepest to
root. So before searching I compute deepest expanded selector when all
items at the bottom.