Add Build.dataSources (#1819)
Ref https://github.com/webstudio-is/webstudio-builder/issues/1529

Here initial setup for data sources which are will be used to maintain
runtime stores. It has unique id, name for ui, type and defaultValue to
initialize store with it.

For now added only string and boolean types which are necessary for Form
component. In the future will be extended with more types.