Replace sharp with image-size (#2303)
Sharp is image processing library. Though we use it only to detect
format and size of images. Proocessing is done on build step or on
request by vercel or cloudflare.

Here I replaced sharp with smaller and simpler alternative designed
specifically for format and size detection. Potentially this library can
be used for uploading in worker.

https://github.com/image-size/image-size

```diff
du -ckd1 node_modules
- 680068
+ 659988
Packages
- 1983
+ 1962
```