fix: Handle empty input values from the share project popover (#1978)
fixes #1966 

The steps to reproduce are in the linked issue. The popover shouldn't
close when the name in the input field is empty.

@taylornowotny i would like you to check this one for the error state
please, the Input field didn't have a red outline for error as defined
in the figma file. I added it now, So, in order to check the effects of
this in other place. It would be great if you can check it once.

![Screenshot 2023-07-25 at 17 35
51](https://github.com/webstudio-is/webstudio-builder/assets/11075561/db8087b5-13f0-401d-8874-a7c9e248b350)

---------

Co-authored-by: Oleg Isonen <oleg008@gmail.com>